run bash tests, not zsh
- user might not have `zsh` installed
