+++ code/executor/src/command/bash.rs
@@ -24,3 +24,35 @@ impl Command for Bash {
         Ok(output)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::{command::Command, Ctx};
+
+    #[tokio::test]
+    async fn test_oneline() -> anyhow::Result<()> {
+        let exec = Ctx::default();
+        let cmd = super::Bash;
+
+        let output = cmd.execute(exec, "echo hello there").await?;
+
+        assert_eq!(output, "hello there");
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_multiline() -> anyhow::Result<()> {
+        let exec = Ctx::default();
+        let cmd = super::Bash;
+
+        let input = r#"echo hello
+        echo there"#;
+
+        let output = cmd.execute(exec, input).await?;
+
+        assert_eq!(output, "hello\nthere");
+
+        Ok(())
+    }
+}
