+++ code/executor/src/command/zsh.rs
@@ -24,35 +24,3 @@ impl Command for Zsh {
         Ok(output)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::{command::Command, Ctx};
-
-    #[tokio::test]
-    async fn test_oneline() -> anyhow::Result<()> {
-        let exec = Ctx::default();
-        let cmd = super::Zsh;
-
-        let output = cmd.execute(exec, "echo hello there").await?;
-
-        assert_eq!(output, "hello there");
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_multiline() -> anyhow::Result<()> {
-        let exec = Ctx::default();
-        let cmd = super::Zsh;
-
-        let input = r#"echo hello
-        echo there"#;
-
-        let output = cmd.execute(exec, input).await?;
-
-        assert_eq!(output, "hello\nthere");
-
-        Ok(())
-    }
-}
